apps/ciphercore_ai_terminal/requirements.txt
customtkinter
pygame
python-dotenv
google-generativeai
